Get shapes async
