+++ buildingmotif-app/src/app/app-routing.module.ts
@@ -10,13 +10,12 @@ import { ModelDetailResolver } from '../app/model-detail/model-detail.resolver'
 import { TemplateEvaluateComponent} from './template-evaluate/template-evaluate.component'
 import { TemplateEvaluateResolver} from './template-evaluate/template-evaluate.resolver'
 import { ModelNewComponent } from '../app/model-new/model-new.component'
-import { ModelValidateResolver } from '../app/model-validate/model-validate.resolver'
 
 const routes: Routes = [
   { path: 'models/new', component: ModelNewComponent},
   { path: 'templates/:id', component: TemplateDetailComponent },
   { path: 'templates/:id/evaluate', component: TemplateEvaluateComponent, resolve: {TemplateEvaluateResolver}},
-  { path: 'models/:id', component: ModelDetailComponent, resolve: {ModelDetailResolver, ModelValidateResolver}},
+  { path: 'models/:id', component: ModelDetailComponent, resolve: {ModelDetailResolver}},
   { path: 'templates', component: TemplateSearchComponent, resolve: {templateSearch:TemplateSearchResolver}},
   { path: 'models', component: ModelSearchComponent, resolve: {ModelSearchResolver}},
   { path: '',   redirectTo: '/templates', pathMatch: 'full' },
